app/model-api/requirements.txt
numpy==
Flask==1.1.2
scikit-learn==0.24.1
pyyaml==5.1
xgboost==1.3.0
matplotlib
lightgbm
seaborn